+++ x/stakeibc/keeper/icacallbacks.go
@@ -1,30 +1,17 @@
 package keeper
 
 import (
-	"fmt"
-
-	"github.com/spf13/cast"
-
-	"github.com/golang/protobuf/proto"
-
-	epochtypes "github.com/Stride-Labs/stride/x/epochs/types"
-
 	icacallbackstypes "github.com/Stride-Labs/stride/x/icacallbacks/types"
-	recordstypes "github.com/Stride-Labs/stride/x/records/types"
-	"github.com/Stride-Labs/stride/x/stakeibc/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
 )
-
-// ___________________________________________________________________________________________________
 
 var DELEGATE = "delegate"
 var CLAIM = "claim"
 var UNDELEGATE = "undelegate"
 
-// ICACallbacks wrapper struct for interchainstaking keeper
+// ICACallbacks wrapper struct for stakeibc keeper
 type ICACallback func(Keeper, sdk.Context, channeltypes.Packet, *channeltypes.Acknowledgement_Result, []byte) error
 
 type ICACallbacks struct {
@@ -38,7 +25,6 @@ func (k Keeper) ICACallbackHandler() ICACallbacks {
 	return ICACallbacks{k, make(map[string]ICACallback)}
 }
 
-//callback handler
 func (c ICACallbacks) CallICACallback(ctx sdk.Context, id string, packet channeltypes.Packet, ack *channeltypes.Acknowledgement_Result, args []byte) error {
 	return c.icacallbacks[id](c.k, ctx, packet, ack, args)
 }
@@ -58,172 +44,4 @@ func (c ICACallbacks) RegisterICACallbacks() icacallbackstypes.ICACallbackHandle
 		AddICACallback(DELEGATE, ICACallback(DelegateCallback)).
 		AddICACallback(CLAIM, ICACallback(ReinvestCallback)).
 		AddICACallback(UNDELEGATE, ICACallback(UndelegateCallback))
-}
-
-// ----------------------------------- Delegate Callback ----------------------------------- //
-func (k Keeper) MarshalDelegateCallbackArgs(ctx sdk.Context, delegateCallback types.DelegateCallback) ([]byte, error) {
-	out, err := proto.Marshal(&delegateCallback)
-	if err != nil {
-		k.Logger(ctx).Error(fmt.Sprintf("MarshalDelegateCallbackArgs %v", err.Error()))
-		return nil, err
-	}
-	return out, nil
-}
-
-func (k Keeper) UnmarshalDelegateCallbackArgs(ctx sdk.Context, delegateCallback []byte) (*types.DelegateCallback, error) {
-	unmarshalledDelegateCallback := types.DelegateCallback{}
-	if err := proto.Unmarshal(delegateCallback, &unmarshalledDelegateCallback); err != nil {
-		k.Logger(ctx).Error(fmt.Sprintf("UnmarshalDelegateCallbackArgs %v", err.Error()))
-		return nil, err
-	}
-	return &unmarshalledDelegateCallback, nil
-}
-
-func DelegateCallback(k Keeper, ctx sdk.Context, packet channeltypes.Packet, ack *channeltypes.Acknowledgement_Result, args []byte) error {
-	k.Logger(ctx).Info("DelegateCallback executing", "packet", packet)
-
-	if ack == nil {
-		// transaction on the host chain failed
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "ack is nil, packet %v", packet)
-	}
-
-	// deserialize the args
-	delegateCallback, err := k.UnmarshalDelegateCallbackArgs(ctx, args)
-	if err != nil {
-		return err
-	}
-	k.Logger(ctx).Info(fmt.Sprintf("DelegateCallback %v", delegateCallback))
-	hostZone := delegateCallback.GetHostZoneId()
-	zone, found := k.GetHostZone(ctx, hostZone)
-	if !found {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "host zone not found %s", hostZone)
-	}
-	recordId := delegateCallback.GetDepositRecordId()
-
-	for _, splitDelegation := range delegateCallback.SplitDelegations {
-		amount := cast.ToInt64(splitDelegation.Amount)
-		validator := splitDelegation.Validator
-
-		k.Logger(ctx).Info(fmt.Sprintf("incrementing stakedBal %d on %s", amount, validator))
-		if amount < 0 {
-			errMsg := fmt.Sprintf("Balance to stake was negative: %d", amount)
-			k.Logger(ctx).Error(errMsg)
-			return sdkerrors.Wrapf(sdkerrors.ErrLogic, errMsg)
-		} else {
-			zone.StakedBal += amount
-			success := k.AddDelegationToValidator(ctx, zone, validator, amount)
-			if !success {
-				return sdkerrors.Wrapf(types.ErrValidatorDelegationChg, "Failed to add delegation to validator")
-			}
-			k.SetHostZone(ctx, zone)
-		}
-	}
-
-	k.RecordsKeeper.RemoveDepositRecord(ctx, cast.ToUint64(recordId))
-	return nil
-}
-
-// ----------------------------------- claim callback ----------------------------------- //
-func (k Keeper) MarshalClaimCallbackArgs(ctx sdk.Context, claimCallback types.ClaimCallback) ([]byte, error) {
-	out, err := proto.Marshal(&claimCallback)
-	if err != nil {
-		k.Logger(ctx).Error(fmt.Sprintf("MarshalClaimCallbackArgs %v", err.Error()))
-		return nil, err
-	}
-	return out, nil
-}
-
-func (k Keeper) UnmarshalClaimCallbackArgs(ctx sdk.Context, claimCallback []byte) (*types.ClaimCallback, error) {
-	unmarshalledDelegateCallback := types.ClaimCallback{}
-	if err := proto.Unmarshal(claimCallback, &unmarshalledDelegateCallback); err != nil {
-        k.Logger(ctx).Error(fmt.Sprintf("UnmarshalClaimCallbackArgs %v", err.Error()))
-		return nil, err
-	}
-	return &unmarshalledDelegateCallback, nil
-}
-
-func ClaimCallback(k Keeper, ctx sdk.Context, packet channeltypes.Packet, ack *channeltypes.Acknowledgement_Result, args []byte) error {
-	k.Logger(ctx).Info("ClaimCallback executing", "packet", packet, "ack", ack, "args", args)
-	// deserialize the args
-	claimCallback, err := k.UnmarshalClaimCallbackArgs(ctx, args)
-	if err != nil {
-		return err
-	}
-	k.Logger(ctx).Info(fmt.Sprintf("ClaimCallback %v", claimCallback))
-	userClaimRecord, found := k.RecordsKeeper.GetUserRedemptionRecord(ctx, claimCallback.GetUserRedemptionRecordId())
-	if !found {
-		return sdkerrors.Wrapf(types.ErrRecordNotFound, "user redemption record not found %s", claimCallback.GetUserRedemptionRecordId())
-	}
-
-	if ack == nil {
-		// transaction on the host chain failed
-		// set UserClaimRecord as claimable
-		userClaimRecord.IsClaimable = true
-		k.RecordsKeeper.SetUserRedemptionRecord(ctx, userClaimRecord)
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "ack is nil, tx failed on host chain")
-	}
-	// claim successfully processed
-	k.RecordsKeeper.RemoveUserRedemptionRecord(ctx, claimCallback.GetUserRedemptionRecordId())
-	return nil
-}
-
-// ----------------------------------- Reinvest Callback ----------------------------------- //
-func (k Keeper) MarshalReinvestCallbackArgs(ctx sdk.Context, reinvestCallback types.ReinvestCallback) ([]byte, error) {
-	out, err := proto.Marshal(&reinvestCallback)
-	if err != nil {
-		k.Logger(ctx).Error(fmt.Sprintf("MarshalReinvestCallbackArgs %v", err.Error()))
-		return nil, err
-	}
-	return out, nil
-}
-
-func (k Keeper) UnmarshalReinvestCallbackArgs(ctx sdk.Context, reinvestCallback []byte) (*types.ReinvestCallback, error) {
-	unmarshalledReinvestCallback := types.ReinvestCallback{}
-	if err := proto.Unmarshal(reinvestCallback, &unmarshalledReinvestCallback); err != nil {
-        k.Logger(ctx).Error(fmt.Sprintf("UnmarshalReinvestCallbackArgs %s", err.Error()))
-		return nil, err
-	}
-	return &unmarshalledReinvestCallback, nil
-}
-
-func ReinvestCallback(k Keeper, ctx sdk.Context, packet channeltypes.Packet, ack *channeltypes.Acknowledgement_Result, args []byte) error {
-	k.Logger(ctx).Info("ReinvestCallback executing", "packet", packet)
-
-	if ack == nil {
-		// transaction on the host chain failed
-		// don't create a record
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest, "ack is nil, packet %v", packet)
-	}
-
-	// deserialize the args
-	reinvestCallback, err := k.UnmarshalReinvestCallbackArgs(ctx, args)
-	if err != nil {
-		return err
-	}
-	amount := reinvestCallback.ReinvestAmount.Amount
-	denom := reinvestCallback.ReinvestAmount.Denom
-	
-	// fetch epoch
-	strideEpochTracker, found := k.GetEpochTracker(ctx, epochtypes.STRIDE_EPOCH)
-	if !found {
-		k.Logger(ctx).Error("failed to find epoch")
-		return sdkerrors.Wrapf(types.ErrInvalidLengthEpochTracker, "no number for epoch (%s)", epochtypes.STRIDE_EPOCH)
-	}
-	epochNumber := strideEpochTracker.EpochNumber
-	// create a new record so that rewards are reinvested
-	amt, err := cast.ToInt64E(amount)
-	if err != nil {
-		k.Logger(ctx).Error(fmt.Sprintf("failed to convert amount %v", err.Error()))
-		return err
-	}
-	record := recordstypes.DepositRecord{
-		Amount:             amt,
-		Denom:              denom,
-		HostZoneId:         reinvestCallback.HostZoneId,
-		Status:             recordstypes.DepositRecord_STAKE,
-		Source:             recordstypes.DepositRecord_WITHDRAWAL_ICA,
-		DepositEpochNumber: epochNumber,
-	}
-	k.RecordsKeeper.AppendDepositRecord(ctx, record)
-	return nil
 }
